+++ sdk/storage/storage-common/src/indexPlatform.ts
@@ -31,6 +31,7 @@ export {
 } from "./policies/SharedKeySigning.js";
 export * from "./policies/StorageRedirectRangeHeaderPolicy.js";
 export * from "./policies/StorageRequestFailureDetailsParserPolicy.js";
+export * from "./policies/StorageDataLocalityPolicy.js";
 export * from "#platform/credentials/UserDelegationKeyCredential";
 export type { UserDelegationKey } from "./credentials/UserDelegationKey.js";